cabinet_dentaire_backend/gateway/requirements.txt
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
PyJWT==2.12.1
python-consul==1.1.0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
